# Conflicts:
#	lang/php/lib/Avro.php
#	lang/php/lib/AvroUtil.php
#	lang/php/lib/Datum/AvroIODatumReader.php
#	lang/php/lib/Schema/AvroEnumSchema.php
#	lang/php/lib/Schema/AvroField.php
#	lang/php/lib/Schema/AvroFixedSchema.php
#	lang/php/lib/Schema/AvroNamedSchema.php
#	lang/php/lib/Schema/AvroRecordSchema.php
#	lang/php/lib/Schema/AvroSchema.php
#	lang/php/test/IODatumReaderTest.php
#	lang/php/test/SchemaTest.php

+++ lang/java/protobuf/src/main/java/org/apache/avro/protobuf/ProtoConversions.java
@@ -117,12 +117,12 @@ public class ProtoConversions {
 
     switch (precise) {
     case Millis:
-      seconds = Math.floorDiv(elapsedSinceEpoch, THOUSAND);
-      nanos = (int) Math.floorMod(elapsedSinceEpoch, THOUSAND) * MILLION;
+      seconds = Math.floorDiv(elapsedSinceEpoch, (long) THOUSAND);
+      nanos = (int) Math.floorMod(elapsedSinceEpoch, (long) THOUSAND) * MILLION;
       break;
     case Micros:
-      seconds = Math.floorDiv(elapsedSinceEpoch, MILLION);
-      nanos = (int) Math.floorMod(elapsedSinceEpoch, MILLION) * THOUSAND;
+      seconds = Math.floorDiv(elapsedSinceEpoch, (long) MILLION);
+      nanos = (int) Math.floorMod(elapsedSinceEpoch, (long) MILLION) * THOUSAND;
       break;
     }
